Added some more TIV fancy.  requestless TIVs are displayed separate and differently.  If a TIV has no request then in order to 'verify' it you must type in a quantity > -1.  If a TIV has a request it is displayed with other TIVs from the same request

+++ matrr/forms.py
@@ -327,7 +327,7 @@ class TissueInventoryVerificationForm(Form):
 	primarykey = IntegerField(widget=HiddenInput(), required=False)
 	freezer = CharField(max_length=100, required=False)
 	location = CharField(max_length=100, required=False)
-	quantity = DecimalField(required=False)
+	quantity = FloatField(required=False)
 	units = ChoiceField(choices=Units, required=False)
 	details = CharField(widget=widgets.Textarea(attrs={'cols': 40, 'rows': 2, 'style':"width:100%;",}), required=False)
 	inventory = ChoiceField(choices=InventoryStatus, required=False)
